Cleanup protocol packages for persistence model v2 (#1085)
