--- legend-engine-xt-persistence-protocol/src/main/java/org/finos/legend/engine/protocol/pure/v1/model/packageableElement/persistence/service/output/dataset/deduplication/DeduplicationVisitor.java
@@ -1,24 +1,0 @@
-// Copyright 2022 Goldman Sachs
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.service.output.dataset.deduplication;
-
-public interface DeduplicationVisitor<T>
-{
-    T visitNoDeduplication(NoDeduplication val);
-
-    T visitAnyVersion(AnyVersion val);
-
-    T visitMaxVersion(MaxVersion val);
-}
